test(integration): remove invalid messages conflict setup
